+++ core/src/main/java/org/keycloak/representations/idm/OpenIdFederationRepresentation.java
@@ -1,17 +1,12 @@
 package org.keycloak.representations.idm;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 public class OpenIdFederationRepresentation {
 
     private String internalId;
     private String trustAnchor;
-    private List<String> clientRegistrationTypesSupported;
-
-    private List<String> entityTypes;
-
     private Map<String, String> idpConfiguration  = new HashMap<>();
 
     public OpenIdFederationRepresentation(){}
@@ -32,21 +27,6 @@ public class OpenIdFederationRepresentation {
         this.trustAnchor = trustAnchor;
     }
 
-    public List<String> getClientRegistrationTypesSupported() {
-        return clientRegistrationTypesSupported;
-    }
-
-    public void setClientRegistrationTypesSupported(List<String> clientRegistrationTypesSupported) {
-        this.clientRegistrationTypesSupported = clientRegistrationTypesSupported;
-    }
-
-    public List<String> getEntityTypes() {
-        return entityTypes;
-    }
-
-    public void setEntityTypes(List<String> entityTypes) {
-        this.entityTypes = entityTypes;
-    }
 
     public Map<String, String> getIdpConfiguration() {
         return idpConfiguration;