+++ core/src/main/java/org/keycloak/representations/idm/RealmRepresentation.java
@@ -163,6 +163,9 @@ public class RealmRepresentation {
     private String openIdFederationResolveEndpoint;
     private String openIdFederationHistoricalKeysEndpoint;
     private List<OpenIdFederationRepresentation> openIdFederationList;
+    private List<String> openIdFederationEntityTypes;
+    private List<String> openIdFederationOPClientRegistrationTypesSupported;
+    private List<String> openIdFederationRPClientRegistrationTypesSupported;
 
     // Client Policies/Profiles
 
@@ -1316,6 +1319,30 @@ public class RealmRepresentation {
 
     public void setOpenIdFederationList(List<OpenIdFederationRepresentation> openIdFederationList) {
         this.openIdFederationList = openIdFederationList;
+    }
+
+    public List<String> getOpenIdFederationEntityTypes() {
+        return openIdFederationEntityTypes;
+    }
+
+    public void setOpenIdFederationEntityTypes(List<String> openIdFederationEntityTypes) {
+        this.openIdFederationEntityTypes = openIdFederationEntityTypes;
+    }
+
+    public List<String> getOpenIdFederationOPClientRegistrationTypesSupported() {
+        return openIdFederationOPClientRegistrationTypesSupported;
+    }
+
+    public void setOpenIdFederationOPClientRegistrationTypesSupported(List<String> openIdFederationOPClientRegistrationTypesSupported) {
+        this.openIdFederationOPClientRegistrationTypesSupported = openIdFederationOPClientRegistrationTypesSupported;
+    }
+
+    public List<String> getOpenIdFederationRPClientRegistrationTypesSupported() {
+        return openIdFederationRPClientRegistrationTypesSupported;
+    }
+
+    public void setOpenIdFederationRPClientRegistrationTypesSupported(List<String> openIdFederationRPClientRegistrationTypesSupported) {
+        this.openIdFederationRPClientRegistrationTypesSupported = openIdFederationRPClientRegistrationTypesSupported;
     }
 
     // Client Policies/Profiles