+++ model/jpa/src/main/java/org/keycloak/models/jpa/RealmAdapter.java
@@ -1059,8 +1059,21 @@ public class RealmAdapter implements LegacyRealmModel, JpaModel<RealmEntity> {
     }
 
     @Override
-    public Stream<OpenIdFederationConfig> getTrustAnchorsBasedOnTypes(EntityTypeEnum entityType, ClientRegistrationTypeEnum clientRegistrationType) {
-        return isOpenIdFederationEnabled() ? getOpenIdFederations().stream().filter(x -> x.getEntityTypes().contains(entityType) && x.getClientRegistrationTypesSupported().contains(clientRegistrationType)) : Stream.empty();
+    public boolean isOpenIdFederationTypeRegistrationSupported(EntityTypeEnum entityType, ClientRegistrationTypeEnum clientRegistrationType) {
+        OpenIdFederationGeneralConfig generalConfig = getOpenIdFederationGeneralConfig();
+        if (generalConfig == null ||  generalConfig.getEntityTypes() == null) {
+            return false;
+        }
+
+        boolean check = generalConfig.getEntityTypes().contains(entityType);
+        if (check) {
+            return EntityTypeEnum.OPENID_PROVIDER.equals(EntityTypeEnum.OPENID_PROVIDER) ? generalConfig.getOpClientRegistrationTypesSupported() != null &&
+                    generalConfig.getOpClientRegistrationTypesSupported().contains(clientRegistrationType) : generalConfig.getRpClientRegistrationTypesSupported() != null &&
+                    generalConfig.getRpClientRegistrationTypesSupported().contains(clientRegistrationType);
+
+        } else {
+            return false;
+        }
     }
 
     @Override
@@ -1082,6 +1095,26 @@ public class RealmAdapter implements LegacyRealmModel, JpaModel<RealmEntity> {
             config.setLifespan(getAttribute(Constants.OPENID_FEDERATION_LIFESPAN, 86400));
             config.setFederationResolveEndpoint(getAttribute(Constants.OPENID_FEDERATION_RESOLVE_ENDPOINT));
             config.setFederationHistoricalKeysEndpoint(getAttribute(Constants.OPENID_FEDERATION_HISTORICAL_KEYS_ENDPOINT));
+            String entityTypes = getAttribute(Constants.OPENID_FEDERATION_ENTITY_TYPES);
+            if (entityTypes != null && !entityTypes.isEmpty()) {
+                config.setEntityTypes(Arrays.asList(entityTypes.split(",")).stream()
+                        .map(EntityTypeEnum::valueOf)
+                        .collect(Collectors.toList()));
+            }
+
+            String opClientRegistrationTypesStr = getAttribute(Constants.OPENID_FEDERATION_OP_CLIENT_REGISTRATION_TYPES_SUPPORTED);
+            if (opClientRegistrationTypesStr != null && !opClientRegistrationTypesStr.isEmpty()) {
+                config.setOpClientRegistrationTypesSupported(Arrays.asList(opClientRegistrationTypesStr.split(",")).stream()
+                        .map(ClientRegistrationTypeEnum::valueOf)
+                        .collect(Collectors.toList()));
+            }
+
+            String rpClientRegistrationTypesStr = getAttribute(Constants.OPENID_FEDERATION_RP_CLIENT_REGISTRATION_TYPES_SUPPORTED);
+            if (rpClientRegistrationTypesStr != null && !rpClientRegistrationTypesStr.isEmpty()) {
+                config.setRpClientRegistrationTypesSupported(Arrays.asList(rpClientRegistrationTypesStr.split(",")).stream()
+                        .map(ClientRegistrationTypeEnum::valueOf)
+                        .collect(Collectors.toList()));
+            }
 
             config.setOpenIdFederationList(getOpenIdFederationList(realm.getOpenIdFederationList()));
             return config;
@@ -1113,6 +1146,22 @@ public class RealmAdapter implements LegacyRealmModel, JpaModel<RealmEntity> {
             setAttribute(Constants.OPENID_FEDERATION_LIFESPAN, generalConfig.getLifespan() == null ? String.valueOf(86400) : String.valueOf(generalConfig.getLifespan()));
             setAttribute(Constants.OPENID_FEDERATION_RESOLVE_ENDPOINT, generalConfig.getFederationResolveEndpoint());
             setAttribute(Constants.OPENID_FEDERATION_HISTORICAL_KEYS_ENDPOINT, generalConfig.getFederationHistoricalKeysEndpoint());
+            if (generalConfig.getEntityTypes() == null || generalConfig.getEntityTypes().isEmpty()) {
+                removeAttribute(Constants.OPENID_FEDERATION_ENTITY_TYPES);
+            } else {
+                setAttribute(Constants.OPENID_FEDERATION_ENTITY_TYPES, generalConfig.getEntityTypes().stream().map(EntityTypeEnum::name).collect(Collectors.joining(",")));
+            }
+            if (generalConfig.getOpClientRegistrationTypesSupported() == null || generalConfig.getOpClientRegistrationTypesSupported().isEmpty()) {
+                removeAttribute(Constants.OPENID_FEDERATION_OP_CLIENT_REGISTRATION_TYPES_SUPPORTED);
+            } else {
+                setAttribute(Constants.OPENID_FEDERATION_OP_CLIENT_REGISTRATION_TYPES_SUPPORTED, generalConfig.getOpClientRegistrationTypesSupported().stream().map(ClientRegistrationTypeEnum::name).collect(Collectors.joining(",")));
+            }
+            if (generalConfig.getRpClientRegistrationTypesSupported() == null || generalConfig.getRpClientRegistrationTypesSupported().isEmpty()) {
+                removeAttribute(Constants.OPENID_FEDERATION_RP_CLIENT_REGISTRATION_TYPES_SUPPORTED);
+            } else {
+                setAttribute(Constants.OPENID_FEDERATION_RP_CLIENT_REGISTRATION_TYPES_SUPPORTED, generalConfig.getRpClientRegistrationTypesSupported().stream().map(ClientRegistrationTypeEnum::name).collect(Collectors.joining(",")));
+            }
+
             if (!generalConfig.getOpenIdFederationList().isEmpty()) {
                 realm.setOpenIdFederationList(generalConfig.getOpenIdFederationList().stream().map(fedConfig -> {
                     OpenIdFederationEntity fedEntity = new OpenIdFederationEntity();
@@ -1134,12 +1183,6 @@ public class RealmAdapter implements LegacyRealmModel, JpaModel<RealmEntity> {
             OpenIdFederationConfig fedConfig = new OpenIdFederationConfig();
             fedConfig.setInternalId(fedEntity.getInternalId());
             fedConfig.setTrustAnchor(fedEntity.getTrustAnchor());
-            String clientRegistrationTypesSupportedStr = fedEntity.getConfig().get(Constants.OPENID_FEDERATION_CLIENT_REGISTRATION_TYPES_SUPPORTED);
-            List<String> clientRegistrationTypesSupported = (clientRegistrationTypesSupportedStr == null || clientRegistrationTypesSupportedStr.isEmpty()) ? new ArrayList<>() : Arrays.asList(clientRegistrationTypesSupportedStr.split("##"));
-            fedConfig.setClientRegistrationTypesSupported(clientRegistrationTypesSupported.stream().map(x -> ClientRegistrationTypeEnum.valueOf(x)).collect(Collectors.toList()));
-            String entityTypesStr = fedEntity.getConfig().get(Constants.OPENID_FEDERATION_ENTITY_TYPES);
-            List<String> entityTypes = (entityTypesStr == null || entityTypesStr.isEmpty()) ? new ArrayList<>() : Arrays.asList(entityTypesStr.split("##"));
-            fedConfig.setEntityTypes(entityTypes.stream().map(x -> EntityTypeEnum.valueOf(x)).collect(Collectors.toList()));
             fedConfig.setIdpConfiguration(fedEntity.getIdpConfiguration());
             return fedConfig;
         }).collect(Collectors.toList());
@@ -1164,8 +1207,6 @@ public class RealmAdapter implements LegacyRealmModel, JpaModel<RealmEntity> {
 
     private void transformToOpenIdFederationEntity(OpenIdFederationEntity fedEntity, OpenIdFederationConfig fedConfig){
         fedEntity.setTrustAnchor(fedConfig.getTrustAnchor());
-        fedEntity.getConfig().put(Constants.OPENID_FEDERATION_ENTITY_TYPES, fedConfig.getEntityTypes().stream().map(x -> x.name()).collect(Collectors.joining("##")));
-        fedEntity.getConfig().put(Constants.OPENID_FEDERATION_CLIENT_REGISTRATION_TYPES_SUPPORTED, fedConfig.getClientRegistrationTypesSupported().stream().map(x -> x.name()).collect(Collectors.joining("##")));
         fedEntity.setIdpConfiguration(fedConfig.getIdpConfiguration());
         fedEntity.setRealm(realm);
     }