+++ model/legacy-private/src/main/java/org/keycloak/storage/datastore/LegacyExportImportManager.java
@@ -1280,6 +1280,16 @@ public class LegacyExportImportManager implements ExportImportManager {
             config.setLifespan(rep.getOpenIdFederationLifespan());
             config.setFederationResolveEndpoint(rep.getOpenIdFederationResolveEndpoint());
             config.setFederationHistoricalKeysEndpoint(rep.getOpenIdFederationHistoricalKeysEndpoint());
+            if (rep.getOpenIdFederationEntityTypes() != null) {
+                config.setEntityTypes(rep.getOpenIdFederationEntityTypes().stream().map(EntityTypeEnum::valueOf).collect(Collectors.toList()));
+            }
+            if (rep.getOpenIdFederationOPClientRegistrationTypesSupported() != null) {
+                config.setOpClientRegistrationTypesSupported(rep.getOpenIdFederationOPClientRegistrationTypesSupported().stream().map(ClientRegistrationTypeEnum::valueOf).collect(Collectors.toList()));
+            }
+            if (rep.getOpenIdFederationRPClientRegistrationTypesSupported() != null) {
+                config.setRpClientRegistrationTypesSupported(rep.getOpenIdFederationRPClientRegistrationTypesSupported().stream().map(ClientRegistrationTypeEnum::valueOf).collect(Collectors.toList()));
+            }
+
             if (withFederationList && rep.getOpenIdFederationList() != null && !rep.getOpenIdFederationList().isEmpty()) {
                 config.setOpenIdFederationList(rep.getOpenIdFederationList().stream().map(fedRep -> toModel(fedRep)).collect(Collectors.toList()));
             }
@@ -1293,8 +1303,6 @@ public class LegacyExportImportManager implements ExportImportManager {
         OpenIdFederationConfig fedConfig = new OpenIdFederationConfig();
         fedConfig.setInternalId(representation.getInternalId());
         fedConfig.setTrustAnchor(representation.getTrustAnchor());
-        fedConfig.setEntityTypes(representation.getEntityTypes().stream().map(EntityTypeEnum::valueOf).collect(Collectors.toList()));
-        fedConfig.setClientRegistrationTypesSupported(representation.getClientRegistrationTypesSupported().stream().map(ClientRegistrationTypeEnum::valueOf).collect(Collectors.toList()));
         fedConfig.setIdpConfiguration(representation.getIdpConfiguration());
         return fedConfig;
     }