+++ server-spi/src/main/java/org/keycloak/models/OpenIdFederationConfig.java
@@ -1,19 +1,12 @@
 package org.keycloak.models;
 
-import org.keycloak.models.enums.ClientRegistrationTypeEnum;
-import org.keycloak.models.enums.EntityTypeEnum;
-
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 public class OpenIdFederationConfig {
 
     private String internalId;
     private String trustAnchor;
-    private List<ClientRegistrationTypeEnum> clientRegistrationTypesSupported;
-
-    private List<EntityTypeEnum> entityTypes;
     private Map<String, String> idpConfiguration  = new HashMap<>();
 
     public OpenIdFederationConfig() {}
@@ -34,21 +27,6 @@ public class OpenIdFederationConfig {
         this.trustAnchor = trustAnchor;
     }
 
-    public List<ClientRegistrationTypeEnum> getClientRegistrationTypesSupported() {
-        return clientRegistrationTypesSupported;
-    }
-
-    public void setClientRegistrationTypesSupported(List<ClientRegistrationTypeEnum> clientRegistrationTypesSupported) {
-        this.clientRegistrationTypesSupported = clientRegistrationTypesSupported;
-    }
-
-    public List<EntityTypeEnum> getEntityTypes() {
-        return entityTypes;
-    }
-
-    public void setEntityTypes(List<EntityTypeEnum> entityTypes) {
-        this.entityTypes = entityTypes;
-    }
 
     public Map<String, String> getIdpConfiguration() {
         return idpConfiguration;