+++ server-spi/src/main/java/org/keycloak/models/RealmModel.java
@@ -329,12 +329,8 @@ public interface RealmModel extends RoleContainerModel {
         return getOpenIdFederationGeneralConfig() != null;
     };
 
-    default Stream<OpenIdFederationConfig> getTrustAnchorsBasedOnTypes(EntityTypeEnum entityType, ClientRegistrationTypeEnum clientRegistrationType) {
-        return Stream.empty();
-    };
-
-    default Set<String> getTrustAnchorsIdsBasedOnTypes(EntityTypeEnum entityType, ClientRegistrationTypeEnum clientRegistrationType) {
-        return getTrustAnchorsBasedOnTypes(entityType, clientRegistrationType).map(OpenIdFederationConfig::getTrustAnchor).collect(Collectors.toSet());
+    default boolean isOpenIdFederationTypeRegistrationSupported(EntityTypeEnum entityType, ClientRegistrationTypeEnum clientRegistrationType) {
+        return isOpenIdFederationEnabled() && getOpenIdFederationGeneralConfig().getEntityTypes() != null  && getOpenIdFederationGeneralConfig().getEntityTypes().contains(entityType) && (EntityTypeEnum.OPENID_RELYING_PARTY.equals(entityType) ? getOpenIdFederationGeneralConfig().getOpClientRegistrationTypesSupported().contains(clientRegistrationType) : getOpenIdFederationGeneralConfig().getRpClientRegistrationTypesSupported().contains(clientRegistrationType)) ;
     };
 
     default void setOpenIdFederationGeneralConfig(OpenIdFederationGeneralConfig generalConfig) {