+++ services/src/main/java/org/keycloak/protocol/oidc/endpoints/request/AuthorizationEndpointRequestParserProcessor.java
@@ -109,7 +109,7 @@ public class AuthorizationEndpointRequestParserProcessor {
         }
     }
 
-    public static AuthorizationEndpointRequest parseRequestOpenIdFederation(EventBuilder event, KeycloakSession session, MultivaluedMap<String, String> requestParams, EndpointType endpointType) {
+    public static AuthorizationEndpointRequest parseRequestOpenIdFederation(EventBuilder event, KeycloakSession session, ClientModel client, MultivaluedMap<String, String> requestParams, EndpointType endpointType) {
         try {
             AuthorizationEndpointRequest request = parseRequestCommon(requestParams, endpointType);
 
@@ -118,12 +118,27 @@ public class AuthorizationEndpointRequestParserProcessor {
 
             if (requestParam != null && requestUriParam != null) {
                 throw new RuntimeException("Illegal to use both 'request' and 'request_uri' parameters together");
+            } else  if (requestParam == null && requestUriParam == null) {
+                throw new RuntimeException("For automatic client registration");
             }
 
+            String requestObjectRequired = OIDCAdvancedConfigWrapper.fromClientModel(client).getRequestObjectRequired();
+
+            if (OIDCConfigAttributes.REQUEST_OBJECT_REQUIRED_REQUEST.equals(requestObjectRequired)
+                    && requestParam == null) {
+                throw new RuntimeException("Client is required to use 'request' parameter.");
+            }
+//            else if (OIDCConfigAttributes.REQUEST_OBJECT_REQUIRED_REQUEST_URI.equals(requestObjectRequired)
+//                    && requestUriParam == null) {
+//                throw new RuntimeException("Client is required to use 'request_uri' parameter.");
+            //            }
+
             if (requestParam != null) {
-                new OpenIdFederationAuthzEndpointRequestObjectParser(session, requestParam).parseRequest(request);
+                new OpenIdFederationAuthzEndpointRequestObjectParser(session, requestParam, client).parseRequest(request);
             } else if (requestUriParam != null) {
                 throw new RuntimeException("requestUriParam not supported for automatic client registration");
+            } else {
+                throw new RuntimeException("For automatic registration is required to use 'request' or 'request_uri' parameter.");
             }
 
             if (Profile.isFeatureEnabled(Profile.Feature.DYNAMIC_SCOPES)) {
@@ -133,9 +148,12 @@ public class AuthorizationEndpointRequestParserProcessor {
             return request;
 
         } catch (Exception e) {
+            client.setEnabled(false);
+            client.updateClient();
             ServicesLogger.LOGGER.invalidRequest(e);
             event.error(Errors.INVALID_REQUEST);
-            throw new ErrorPageException(session, Response.Status.BAD_REQUEST, Messages.INVALID_REQUEST);
+            event.detail("Automatic","true");
+            throw new ErrorPageException(session, Response.Status.BAD_REQUEST, Messages.INVALID_REQUEST_AUTOMATIC_REGISTRATION);
         }
     }