+++ services/src/main/java/org/keycloak/protocol/oidc/federation/OpenIdFederationAuthzEndpointRequestObjectParser.java
@@ -1,9 +1,16 @@
 package org.keycloak.protocol.oidc.federation;
 
 import com.fasterxml.jackson.databind.JsonNode;
+import jakarta.ws.rs.core.Response;
+import org.keycloak.events.Errors;
+import org.keycloak.models.ClientModel;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.protocol.oidc.OIDCLoginProtocol;
 import org.keycloak.protocol.oidc.endpoints.request.AuthzEndpointRequestParser;
+import org.keycloak.services.ErrorPageException;
+import org.keycloak.services.Urls;
+import org.keycloak.services.messages.Messages;
+import org.keycloak.urls.UrlType;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -12,8 +19,8 @@ public class OpenIdFederationAuthzEndpointRequestObjectParser extends AuthzEndpo
 
     private final JsonNode requestParams;
 
-    public OpenIdFederationAuthzEndpointRequestObjectParser(KeycloakSession session, String requestObject) {
-        this.requestParams = session.tokens().decode(requestObject, JsonNode.class);
+    public OpenIdFederationAuthzEndpointRequestObjectParser(KeycloakSession session, String requestObject, ClientModel client) {
+        this.requestParams = session.tokens().decodeClientJWT(requestObject, client, createRequestObjectValidator(session), JsonNode.class);
 
         if (this.requestParams == null) {
             throw new RuntimeException("Failed to verify signature on 'request' object");
@@ -21,6 +28,12 @@ public class OpenIdFederationAuthzEndpointRequestObjectParser extends AuthzEndpo
 
         if (requestParams.has(OIDCLoginProtocol.REQUEST_URI_PARAM)) {
             throw new RuntimeException("The request_uri claim should not be set in the request object");
+        }
+
+        if (requestParams.has("sub")  || ! requestParams.has("jti")
+                || !requestParams.has("exp") || !requestParams.has("iss")
+                ||  !Urls.realmIssuer(session.getContext().getUri(UrlType.FRONTEND).getBaseUri(), session.getContext().getRealm().getName()).equals(requestParams.get("aud"))) {
+            throw new RuntimeException(Messages.OPENID_FEDERATION_AUTOMATIC_FALSE_REQUEST_OBJECT);
         }
 
         session.setAttribute(AuthzEndpointRequestParser.AUTHZ_REQUEST_OBJECT, requestParams);