+++ services/src/main/java/org/keycloak/services/clientregistration/openid_federation/OpenIdFederationClientRegistrationService.java
@@ -15,6 +15,7 @@ import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.RealmModel;
 import org.keycloak.models.enums.ClientRegistrationTypeEnum;
 import org.keycloak.models.enums.EntityTypeEnum;
+import org.keycloak.models.OpenIdFederationConfig;
 import org.keycloak.utils.OpenIdFederationTrustChainProcessor;
 import org.keycloak.representations.idm.ClientRepresentation;
 import org.keycloak.representations.openid_federation.EntityStatement;
@@ -29,7 +30,6 @@ import org.keycloak.urls.UrlType;
 import org.keycloak.util.TokenUtil;
 
 import java.net.URI;
-import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -47,8 +47,7 @@ public class OpenIdFederationClientRegistrationService extends AbstractClientReg
     @Consumes({"application/entity-statement+jwt", "application/trust-chain+json"})
     public Response explicitClientRegistration(String body, @Context HttpHeaders headers) {
         RealmModel realm = session.getContext().getRealm();
-        Set<String> trustAnchorIds = realm.getTrustAnchorsIdsBasedOnTypes(EntityTypeEnum.OPENID_PROVIDER, ClientRegistrationTypeEnum.EXPLICIT);
-        if (trustAnchorIds.isEmpty()) {
+        if (!realm.isOpenIdFederationTypeRegistrationSupported(EntityTypeEnum.OPENID_PROVIDER, ClientRegistrationTypeEnum.EXPLICIT) || realm.getOpenIdFederations().isEmpty()) {
             throw new ErrorResponseException(Errors.INVALID_REQUEST, "Explicit OpenID Federation Client Registration is not supported in this realm", Response.Status.BAD_REQUEST);
         }
         checkSsl();
@@ -65,7 +64,7 @@ public class OpenIdFederationClientRegistrationService extends AbstractClientReg
             trustChainProcessor.validationRules(statement, true);
 
             logger.info("starting validating trust chains");
-            TrustChainResolution validChain = trustChainProcessor.constructTrustChains(statement, trustAnchorIds, true);
+            TrustChainResolution validChain = trustChainProcessor.constructTrustChains(statement, realm.getOpenIdFederations().stream().map(OpenIdFederationConfig::getTrustAnchor).collect(Collectors.toSet()), true);
             if (validChain == null) {
                 throw new ErrorResponseException(Errors.INVALID_TRUST_ANCHOR, "No trusted trust anchor could be found", Response.Status.NOT_FOUND);
             }